Final Project/requirements.txt
apache-airflow-providers-apache-spark
apache-airflow-providers-apache-hive
kafka-python
isort
black